requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
aiohttp=3.9.1=pypi_0
aiosignal=1.3.1=pypi_0
anyio=4.6.2.post1=pypi_0
argon2-cffi=23.1.0=pypi_0
argon2-cffi-bindings=21.2.0=pypi_0
arrow=1.3.0=pypi_0
asttokens=2.4.1=pypi_0
async-lru=2.0.4=pypi_0
async-timeout=4.0.3=pypi_0
attrs=23.2.0=pypi_0
babel=2.16.0=pypi_0
beautifulsoup4=4.12.3=pypi_0
bleach=6.1.0=pypi_0
blosc2=2.5.1=pypi_0
ca-certificates=2023.12.12=h06a4308_0
certifi=2023.11.17=pypi_0
cffi=1.17.1=pypi_0
charset-normalizer=3.3.2=pypi_0
click=8.1.7=pypi_0
comm=0.2.2=pypi_0
datasets=2.16.1=pypi_0
debugpy=1.8.7=pypi_0
decorator=5.1.1=pypi_0
defusedxml=0.7.1=pypi_0
dill=0.3.7=pypi_0
exceptiongroup=1.2.2=pypi_0
executing=2.1.0=pypi_0
fastjsonschema=2.20.0=pypi_0
filelock=3.13.1=pypi_0
fqdn=1.5.1=pypi_0
frozenlist=1.4.1=pypi_0
fsspec=2023.10.0=pypi_0
googledrivedownloader=0.4=pypi_0
h11=0.14.0=pypi_0
httpcore=1.0.6=pypi_0
httpx=0.27.2=pypi_0
huggingface-hub=0.24.7=pypi_0
idna=3.6=pypi_0
importlib-metadata=8.5.0=pypi_0
ipykernel=6.29.5=pypi_0
ipython=8.18.1=pypi_0
ipywidgets=8.1.5=pypi_0
isodate=0.6.1=pypi_0
isoduration=20.11.0=pypi_0
jedi=0.19.1=pypi_0
jinja2=3.1.3=pypi_0
joblib=1.3.2=pypi_0
json5=0.9.25=pypi_0
jsonpointer=3.0.0=pypi_0
jsonschema=4.23.0=pypi_0
jsonschema-specifications=2024.10.1=pypi_0
jupyter=1.1.1=pypi_0
jupyter-client=8.6.3=pypi_0
jupyter-console=6.6.3=pypi_0
jupyter-core=5.7.2=pypi_0
jupyter-events=0.10.0=pypi_0
jupyter-lsp=2.2.5=pypi_0
jupyter-server=2.14.2=pypi_0
jupyter-server-terminals=0.5.3=pypi_0
jupyterlab=4.2.5=pypi_0
jupyterlab-pygments=0.3.0=pypi_0
jupyterlab-server=2.27.3=pypi_0
jupyterlab-widgets=3.0.13=pypi_0
ld_impl_linux-64=2.38=h1181459_1
libffi=3.4.4=h6a678d5_0
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libstdcxx-ng=11.2.0=h1234567_1
markupsafe=2.1.3=pypi_0
matplotlib-inline=0.1.7=pypi_0
mistune=3.0.2=pypi_0
msgpack=1.0.8=pypi_0
multidict=6.0.4=pypi_0
multiprocess=0.70.15=pypi_0
nbclient=0.10.0=pypi_0
nbconvert=7.16.4=pypi_0
nbformat=5.10.4=pypi_0
ncurses=6.4=h6a678d5_0
ndindex=1.8=pypi_0
nest-asyncio=1.6.0=pypi_0
networkx=3.2.1=pypi_0
nltk=3.8.1=pypi_0
notebook=7.2.2=pypi_0
notebook-shim=0.2.4=pypi_0
numexpr=2.10.1=pypi_0
numpy=1.26.2=pypi_0
openprompt=1.0.1=pypi_0
openssl=3.0.12=h7f8727e_0
overrides=7.7.0=pypi_0
packaging=23.2=pypi_0
pandas=2.1.4=pypi_0
pandocfilters=1.5.1=pypi_0
parso=0.8.4=pypi_0
pexpect=4.9.0=pypi_0
pillow=10.2.0=pypi_0
pip=23.3.1=py39h06a4308_0
platformdirs=4.3.6=pypi_0
prometheus-client=0.21.0=pypi_0
prompt-toolkit=3.0.48=pypi_0
protobuf=4.25.2=pypi_0
psutil=6.1.0=pypi_0
ptyprocess=0.7.0=pypi_0
pure-eval=0.2.3=pypi_0
py-cpuinfo=9.0.0=pypi_0
pyarrow=14.0.2=pypi_0
pyarrow-hotfix=0.6=pypi_0
pycparser=2.22=pypi_0
pygments=2.18.0=pypi_0
pyparsing=3.1.1=pypi_0
python=3.9.18=h955ad1f_0
python-dateutil=2.8.2=pypi_0
python-json-logger=2.0.7=pypi_0
pytz=2023.3.post1=pypi_0
pyyaml=6.0.1=pypi_0
pyzmq=26.2.0=pypi_0
rdflib=7.0.0=pypi_0
readline=8.2=h5eee18b_0
referencing=0.35.1=pypi_0
regex=2023.12.25=pypi_0
requests=2.31.0=pypi_0
rfc3339-validator=0.1.4=pypi_0
rfc3986-validator=0.1.1=pypi_0
rouge=1.0.0=pypi_0
rpds-py=0.20.0=pypi_0
safetensors=0.4.5=pypi_0
scikit-learn=1.3.2=pypi_0
scipy=1.11.4=pypi_0
send2trash=1.8.3=pypi_0
sentencepiece=0.1.96=pypi_0
setuptools=68.2.2=py39h06a4308_0
six=1.16.0=pypi_0
sniffio=1.3.1=pypi_0
soupsieve=2.6=pypi_0
sqlite=3.41.2=h5eee18b_0
stack-data=0.6.3=pypi_0
tables=3.9.2=pypi_0
tensorboardx=2.6.2.2=pypi_0
terminado=0.18.1=pypi_0
threadpoolctl=3.2.0=pypi_0
tinycss2=1.3.0=pypi_0
tk=8.6.12=h1ccaba5_0
tokenizers=0.12.1=pypi_0
tomli=2.0.2=pypi_0
torch=1.11.0+cu113=pypi_0
torch-geometric=2.2.0=pypi_0
torch-scatter=2.0.9=pypi_0
torch-sparse=0.6.13=pypi_0
torchaudio=0.11.0+cu113=pypi_0
torchvision=0.12.0+cu113=pypi_0
tornado=6.4.1=pypi_0
tqdm=4.66.1=pypi_0
traitlets=5.14.3=pypi_0
transformers=4.33.1=pypi_0
types-python-dateutil=2.9.0.20241003=pypi_0
typing-extensions=4.9.0=pypi_0
tzdata=2023.4=pypi_0
uri-template=1.3.0=pypi_0
urllib3=2.1.0=pypi_0
wcwidth=0.2.13=pypi_0
webcolors=24.8.0=pypi_0
webencodings=0.5.1=pypi_0
websocket-client=1.8.0=pypi_0
wheel=0.41.2=py39h06a4308_0
widgetsnbextension=4.0.13=pypi_0
xxhash=3.4.1=pypi_0
xz=5.4.5=h5eee18b_0
yacs=0.1.8=pypi_0
yarl=1.9.4=pypi_0
zipp=3.20.2=pypi_0
zlib=1.2.13=h5eee18b_0
